requirements.txt
pyobjc-core>=12.0
pyobjc-framework-Cocoa>=12.0
pyobjc-framework-EventKit>=12.0
